+++ src/uiprotect/api.py
@@ -1806,10 +1806,12 @@ class ProtectApiClient(BaseApiClient):
         *,
         volume: int | None = None,
         repeat_times: int | None = None,
+        ringtone_id: str | None = None,
+        track_no: int | None = None,
     ) -> None:
         """Plays chime tones on a chime"""
         data: dict[str, Any] | None = None
-        if volume or repeat_times:
+        if volume or repeat_times or ringtone_id or track_no:
             chime = self.bootstrap.chimes.get(device_id)
             if chime is None:
                 raise BadRequest("Invalid chime ID %s", device_id)
@@ -1817,8 +1819,11 @@ class ProtectApiClient(BaseApiClient):
             data = {
                 "volume": volume or chime.volume,
                 "repeatTimes": repeat_times or chime.repeat_times,
-                "trackNo": chime.track_no,
+                "trackNo": track_no or chime.track_no,
             }
+            if ringtone_id:
+                data["ringtoneId"] = ringtone_id
+                data.pop("trackNo", None)
 
         await self.api_request(
             f"chimes/{device_id}/play-speaker",
@@ -1829,6 +1834,16 @@ class ProtectApiClient(BaseApiClient):
     async def play_buzzer(self, device_id: str) -> None:
         """Plays chime tones on a chime"""
         await self.api_request(f"chimes/{device_id}/play-buzzer", method="post")
+
+    async def set_light_is_led_force_on(
+        self, device_id: str, is_led_force_on: bool
+    ) -> None:
+        """Sets isLedForceOn for light."""  # workaround because forceOn doesnt work via websocket
+        await self.api_request(
+            f"lights/{device_id}",
+            method="patch",
+            json={"lightOnSettings": {"isLedForceOn": is_led_force_on}},
+        )
 
     async def clear_tamper_sensor(self, device_id: str) -> None:
         """Clears tamper status for sensor"""